refactor(server): extract the integration factories to src/composition/
The first extraction against the module map. server.ts 4,275 -> 4,077 lines,
with 311 lines moved into four focused modules and no behaviour change.

A ROUTE CONTRACT TEST COMES FIRST. Decomposing createApp() means moving
registration calls between files, and the failure mode is not a crash -- it is
one registerXRoutes() silently not being called, or two middleware swapping
places. Either produces a green build, a green type-check, and a 404 (or an
unguarded request) nobody notices until a user hits it. So
tests/architecture/routeInventory.test.ts records all 454 routes and 14
middleware layers IN REGISTRATION ORDER (Express matches in that order, so
sorting would hide the reordering this exists to catch) and fails on any change.

Mutation-tested: commenting out registerClockSkewRoutes reports its four routes
by name; swapping caseIdGate with caseLockGate fails the middleware assertion --
that one is the case-id validator and the case-password gate trading places,
which no type-check would catch.

THE LOGGER HAD TO MOVE FIRST. tlsFetchFor warns when an integration's TLS trust
is misconfigured, and warnLine lived in server.ts -- so lifting the factories out
would have created a server -> composition -> server runtime import cycle, the
exact hazard check:imports exists to reject. logging/serverLogger.ts now owns the
mutable binding, so both sides import downward and neither imports the other.
setServerLogger/getServerLogger are re-exported from server.ts, keeping the
eleven test files that reach for them there working against the same binding.

WHAT MOVED, and why these three files:
  - composition/env.ts             isEnvFlag, numEnv, positiveIntEnv. Three
                                   parsers that each exist because the obvious
                                   inline expression got a case wrong.
  - composition/tlsFetch.ts        TLS_HOST_URL + tlsFetchFor, including the
                                   #246 loopback-guard reasoning verbatim.
  - composition/integrationClients.ts
                                   The seven client builders and their option
                                   builders. One contract binds them: read env,
                                   return a client, or return undefined when the
                                   integration is not configured -- undefined is
                                   not an error, it is how an optional
                                   integration stays off and its dashboard
                                   button stays hidden.

buildRuntimePipeline and the AI provider factories deliberately stay put. They
are a different cluster (provider construction, not integration clients) and
moving both at once doubles the review surface for no gain.

NO CALLER CHANGED. Every external user -- five push scripts and three test files
-- imports these from src/server.js, so server.ts re-exports them. The extraction
moved the definitions, not the public surface.

CONTRIBUTING gains two workflow notes learned the hard way here: record the size
ledger LAST (--update refuses to raise, so recording a shrink and then editing
again leaves a number the gate rejects and --update cannot fix), and `npm run
format` is not the gate -- it reformatted all of server.ts into a ~1,000-line
reflow that the gate would have skipped, since the gate leaves never-clean files
alone.

Verified: build, typecheck, lint, format:check, check:size (server.ts re-recorded
at 4,077), check:imports, check:boundaries, and the full suite -- 530 files,
6,692 tests, all passing.

Refs #384
