+++ companion/scripts/file-size-ledger.json
@@ -13,5 +13,5 @@
   "routes/import.ts": 1764,
   "routes/mcp.ts": 845,
   "routes/velociraptor.ts": 888,
-  "server.ts": 4276
+  "server.ts": 4077
 }
